updated issue 15 after merging with development
