+++ test/src/parse_airports_test.c
@@ -11,7 +11,7 @@
 #include<string.h>
 #include <stdbool.h>
 
-#include "../include/data_types.h"
+#include "../../include/data_types.h"
