+++ test/src/test_main.c
@@ -10,6 +10,8 @@
 
 #include "../../include/data_types.h"
 #include "../../include/common_functions.h"
+#include "../../include/helper_functions.h"
+#include "../../include/parse_airports.h"
 #include "../../include/strat_airlift_functions.h"
 #include "../include/strat_airlift_functions_tests.h"
 /**
@@ -26,6 +28,35 @@ struct SIZES *my_sizes;
  * @return 0 if the application executed successfully.  Any other number if their was an error.
  */
 int main() {
+    
+    /* get 2D array of objects */
+
+    int size[7] ;
+
+    char line[N];
+    char filename_1[100]= "airports_empty.csv";
+    char filename_2[100]= "airports_known_number.csv";
+    char filename_3[100] = "airports_copy_wider_size_array.csv";
+        
+    printf("TEST CASE 1: Empty CSV\n");
+    printf("OUTPUT EXPECTED: O airports and returns 0 because it runs effectively\n");
+    struct AIRPORT** arr_1 = get_2D_array(size, line, filename_1);
+    int num_of_airports_1 =parse_airports(arr_1, line);
+    printf("%d\n\n",num_of_airports_1);
+
+    printf("TEST CASE 2: known length of CSV\n");
+    printf("OUTPUT EXPECTED: 5 airports and returns 0 because it runs effectively\n");
+    struct AIRPORT** arr_2 = get_2D_array(size, line,filename_2);
+    int num_of_airports_2 =parse_airports(arr_2, line);
+    printf("%d\n\n",num_of_airports_2);
+
+    printf("TEST CASE 3: Continet size increased to 8\n");
+    printf("OUTPUT EXPECTED: returns -1 as file fails to \n\n");
+    struct AIRPORT** arr_3 = get_2D_array(size, line, filename_3);
+    int num_of_airports_3 =parse_airports(arr_3, line);
+    printf("%d\n\n",num_of_airports_3);
+    
+    
     int rtn_val=0;                                      // The current state of the function to return
 
     printf("=== Testing calc_shortest_path ===\n");
